LPS-60606 Move to the right folder, see modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/portlet/configuration/icon/DocumentTypesPortletConfigurationIconFactory.java
