+++ modules/apps/export-import/export-import-web/src/main/java/com/liferay/exportimport/web/portlet/configuration/icon/ExportImportPortletConfigurationIconFactory.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.staging.configuration.web.configuration.icon;
+package com.liferay.exportimport.web.portlet.configuration.icon;
 
 import com.liferay.portal.kernel.portlet.configuration.icon.BasePortletConfigurationIconFactory;
 import com.liferay.portal.kernel.portlet.configuration.icon.PortletConfigurationIcon;
@@ -26,17 +26,17 @@ import org.osgi.service.component.annotations.Component;
  * @author Eudaldo Alonso
  */
 @Component(immediate = true, service = PortletConfigurationIconFactory.class)
-public class StagingPortletConfigurationIconFactory
+public class ExportImportPortletConfigurationIconFactory
 	extends BasePortletConfigurationIconFactory {
 
 	@Override
 	public PortletConfigurationIcon create(HttpServletRequest request) {
-		return new StagingPortletConfigurationIcon(request);
+		return new ExportImportPortletConfigurationIcon(request);
 	}
 
 	@Override
 	public double getWeight() {
-		return 10.0;
+		return 11.0;
 	}
 
 }
